Create new ProposalTestCase automated test class.
Abstract the creation of a Proposal in a new test case for easier
inheritance and less code duplication.
